simplify delete logic, always goes back to first file now
